+++ apps/tergite-mss/services/auth/providers/dtos.py
@@ -1,5 +1,8 @@
 """Data Transfer Objects for the auth providers"""
+from fastapi import Query
 from pydantic import BaseModel
+
+from utils.models import create_partial_model
 
 
 class AuthProviderRead(BaseModel):
@@ -13,3 +16,20 @@ class AuthProvider(AuthProviderRead):
     """The internal schema for all auth providers"""
 
     email_domain: str
+
+    def as_public_view(self) -> AuthProviderRead:
+        """Returns this instance as a AuthProviderRead instance"""
+        return AuthProviderRead(
+            name=self.name,
+            url=self.url,
+        )
+
+
+# derived modela
+
+AuthProviderQuery = create_partial_model(
+    "AuthProviderQuery",
+    original=AuthProvider,
+    default=Query(None),
+    exclude=("url",),
+)